backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pdf2docx
python-docx
pytest
httpx
authlib
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
aiosqlite
fastapi-mail
itsdangerous
google-auth
